Implement recursive parsing with same environment.
This allows the same lug::environment to be used in multiple,
potentially nested parsing contexts (and thus potentially different
grammars, although it can be used with the same grammar as well).
Specifically, nested parsing is now allowed during the evaluation of
semantic actions once a parsing expression is matched and accepted.

When combined with semantic predicates to control which rules are
applied, this allows different sub-trees within the same
grammar to be re-entered from inside of different evaluation contexts.
This approach has been used to implement function call evaluation in
the BASIC interpreter sample program.

Recursive parsing will also allow parsing conditions and symbol tables
to be implemented in the future on top of this feature.