Implemented parsing conditions.
A parsing condition is a boolean variable that can be used to control
wether a parsing rule matches, independent of the parsing prefix,
allowing for dynamic branching within a grammar. Parsing conditions can
be string identifiers naming a condition within an environment, or they
can be references to C++ `bool` variables with lifetime managed by the
user. Parsing conditions are implemented on top of the existing support
for syntactic predicates.

To support using parsing conditions within a grammar, four new
combinators have been introduced. The `when` and `unless` combinators
are used to test if a parsing condition is true or false, respectively.
The `set` combinator sets parsing condition to true and `unset` sets a
parsing condition to false.