Conformance and stability improvements to BASIC interpreter.
The read_iter_ was being invalidated when adding new DATA elements if
the data_ list was empty. This could sometimes cause a crash when the
next READ was executed. This has been fixed.

Additionally made the CLEAR command clear all data, variables, arrays
tables and functions in addition to the line list.