Change the predicate argument type to environment.
Changing this to the environment from the parser allows predicate
bodies to be defined befored the parser class definition. Furthermore,
it keeps the API of predicates consistent with semantic actions and
syntactic captures, which also take the environment as an argument.

To facilitate this change, the parser instance can now be accessed
from the environment at any point within the scope of parser::parse.